fix: cleaning code
